Add minor code cleanup & styling tweaks
