+++ gitweb/public/less/index.less
@@ -10,6 +10,16 @@
   flex: @val;
 }
 
+.padded {
+  padding: 10px;
+}
+
+.size {
+  @small: 5px;
+  @medium: 10px;
+  @large: 20px;
+}
+
 .fill() {
   width: 100%;
   height: 100%;
@@ -55,3 +65,18 @@ html, body, h1, h2, h3, h4, h5 { margin: 0; padding: 0; border: 0; }
   }
 }
 
+.repo-panel {
+  .padded;
+  .size();
+
+  .repo-panel-header {
+    margin-bottom: @medium;
+
+  }
+}
+
+.pull-panel {
+  .panel-title {
+    display: inline-block;
+  }
+}
